requirements.txt
Flask~=3.1.0
pandas~=2.2.3
inflection~=0.5.1
MarkupSafe~=3.0.2
odfpy~=1.4.1
plotly-cloud~=0.1.0
dash~=3.3.0
matplotlib~=3.10.7
dash_ag_grid~=32.3.2
pyarrow~=22.0.0
pyyaml~=6.0.3
networkx~=3.6
scipy~=1.16.3
